Cont: 심리 상담 가이드 컴포넌트 분리
