+++ src/pages/counseling/Counseling.tsx
@@ -1,12 +1,12 @@
 /** @jsxImportSource @emotion/react */
-import { css, keyframes } from '@emotion/react';
+import { css } from '@emotion/react';
 import variables from '@styles/Variables';
 import { FormEvent, useEffect, useRef, useState } from 'react';
 import fetchGPT from '../../hooks/useGPT';
 import prevIcon from '/img/icon-page-prev.svg';
 import wishIcon from '/img/icon-wish-profile.svg';
 import sendIcon from '/img/icon-send.svg';
-import arrow from '/img/icon-guide-arrow.svg';
+import CounselingGuide from './CounslingGuide';
 
 interface Message {
   sender?: string;
@@ -106,21 +106,6 @@ const Counseling = () => {
   //가이드 영역
   const [step, setStep] = useState<number>(1);
   const [guideVisible, setGuideVisible] = useState(false);
-  const [throttle, setThrottle] = useState(false);
-
-  const handleStep = (e: React.MouseEvent<HTMLButtonElement>) => {
-    if (!throttle) {
-      setThrottle(true);
-      setTimeout(() => {
-        if ((e.target as HTMLButtonElement).classList.contains('skip')) {
-          setStep(4);
-        } else {
-          setStep(step + 1);
-        }
-        setThrottle(false);
-      }, 500);
-    }
-  };
 
   //페이지 진입시 .5초뒤에 가이드 시작
   useEffect(() => {
@@ -141,108 +126,24 @@ const Counseling = () => {
     moveScrollDown();
   }, [messages]);
 
-  //가이드 설명 위치 변경 함수
-  const guideDesc = (step: number, targetStep: number) => {
-    return step === targetStep
-      ? css`
-          display: block;
-          text-align: center;
-          line-height: 2.2rem;
-          width: 100%;
-          font-size: ${variables.size.big};
-          position: absolute;
-          ${targetStep === 1 && 'left: 50%; bottom: 15rem ;transform: translateX(-50%);'}
-          ${targetStep === 2 && 'left: 50%; top:40rem ;transform: translateX(-50%);'}
-          ${targetStep === 3 &&
-          'text-align: right; right: 3.8rem; top: 15rem; & span {font-size: 1.4rem; font-weight: 300; line-height: 3rem;}'}
-        `
-      : css`
-          display: none;
-        `;
-  };
-
-  //가이드 화살표 위치 변경 함수
-  const guideArrowStyles = (step: number) => css`
-    position: absolute;
-    transform: translateX(-50%);
-    ${step === 1 &&
-    css`
-      left: 50%;
-      bottom: 10rem;
-    `}
-    ${step === 2 &&
-    css`
-      transform: scaleY(-1);
-      left: 50%;
-      top: 34rem;
-    `}
-      ${step === 3 &&
-    css`
-      transform: scaleY(-1);
-      right: 3.8rem;
-      top: 8rem;
-    `}
-      ${step !== 1 &&
-    step !== 2 &&
-    step !== 3 &&
-    css`
-      display: none;
-    `};
-  `;
-
   return (
     <>
-      {guideVisible && (
-        <>
-          <button css={stepLayer} onClick={handleStep} type="button">
-            <button className="skip">건너뛰기</button>
-          </button>
-          <div css={[guideDim, guideVisible ? fadeInAni : fadeOutAni]}>
-            <div className="paginationBox">
-              {Array.from({ length: 3 }).map((_, i) => (
-                <div key={i} className={`pagination ${step > i ? 'active' : ''}`} />
-              ))}
-            </div>
-            <div className="guideDescBox">
-              <div css={guideArrowStyles(step)} className="guideArrow">
-                <img src={arrow} alt="화살표아이콘" />
-              </div>
-              <p css={guideDesc(step, 1)}>
-                배우자나 직장, 가족 관계 등 <br />
-                난임에 대한 모든 고민을 위시와 상담해보세요!
-              </p>
-              <p css={guideDesc(step, 2)}>실시간으로 위시와 대화를 나눌 수 있습니다!</p>
+      {guideVisible && <CounselingGuide step={step} setStep={setStep} guideVisible={guideVisible} />}
 
-              <p css={guideDesc(step, 3)}>
-                상담 종료 버튼을 눌러야만
-                <br />
-                다음에도 상담을 이어갈 수 있어요!
-                <br />
-                <span>상담 내용은 상담을 이어가기 위한 용도로만 사용됩니다.</span>
-              </p>
-            </div>
-          </div>
-        </>
-      )}
-
-      <div css={backGroundColor}>
-        <div css={header} className="header">
+      <div css={BackGroundColor}>
+        <div css={Header}>
           <button className="prev" type="button">
-            <span css={hidden}>뒤로가기</span>
+            <span className="hidden">뒤로가기</span>
           </button>
           <h2>심리상담사 위시</h2>
-          <button css={step === 3 && priority} className="end" type="button">
+          <button css={step === 3 && Priority} className="end" type="button">
             종료
           </button>
         </div>
 
-        <div
-          className="scroll-box"
-          ref={scrollBoxRef}
-          css={{ width: '100%', height: 'calc(100% - 146px)', overflow: 'hidden auto' }}
-        >
-          <div css={dateText}>{toDay}</div>
-          <ul css={[messageBox, step === 2 && priority]}>
+        <div className="scroll-box" ref={scrollBoxRef} css={ScrollBox}>
+          <p className="dateText">{toDay}</p>
+          <ul css={[MessageBox, step === 2 && Priority]}>
             {messages.map(({ sender, message }, idx) => (
               <li key={`${sender}-${message!.slice(0, 10)}-${idx}`} className={sender}>
                 <p> {message}</p>
@@ -251,10 +152,10 @@ const Counseling = () => {
           </ul>
         </div>
 
-        <form css={[inputBox, step === 1 && priority]} onSubmit={(e) => handleSubmit(e)}>
+        <form css={[InputBox, step === 1 && Priority]} onSubmit={(e) => handleSubmit(e)}>
           <input type="text" name="" id="" value={userInput} onChange={(e) => setUserInput(e.target.value)} />
           <button>
-            <span css={hidden}>전송</span>
+            <span className="hidden">전송</span>
           </button>
           {isLoading && '로딩중'}
         </form>
@@ -265,113 +166,27 @@ const Counseling = () => {
 
 export default Counseling;
 
-const hidden = css`
-  position: absolute !important;
-  width: 0.1rem;
-  height: 0.1rem;
-  overflow: hidden;
-  clip: rect(0.1rem 0.1rem 0.1rem 0.1rem);
-`;
-
-const fadeIn = keyframes`
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-`;
-
-const fadeOut = keyframes`
-  from {
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-  }
-`;
-const fadeInAni = css`
-  animation: ${fadeIn} 0.5s forwards;
-`;
-
-const fadeOutAni = css`
-  animation: ${fadeOut} 0.5s forwards;
-`;
-
-const guideDim = css`
-  position: absolute;
-  width: 50rem;
-  left: 50%;
-  top: 0;
-  bottom: 0;
-  transform: translateX(-50%);
-  background-color: rgba(0, 0, 0, 0.3);
-  z-index: 2;
-  backdrop-filter: blur(0.3rem);
-  color: ${variables.colors.white};
-
-  .paginationBox {
-    display: flex;
-    width: 100%;
-    height: 2rem;
-    margin-top: 7.2rem;
-    align-items: center;
-    justify-content: center;
-    gap: 1rem;
-
-    .pagination {
-      width: 1rem;
-      height: 1rem;
-      border-radius: 50%;
-      background-color: ${variables.colors.gray50};
-      transition: all ${variables.TransitionDuration};
-    }
-    .pagination.active {
-      background-color: ${variables.colors.white};
-    }
-  }
-`;
-
-const stepLayer = css`
-  position: absolute;
-  width: 50rem;
-  left: 50%;
-  top: 0;
-  bottom: 0;
-  transform: translateX(-50%);
-  z-index: 3;
-  color: ${variables.colors.white};
-
-  .skip {
-    position: absolute;
-    left: 1.8rem;
-    top: 7.2rem;
-    z-index: 5;
-    font-size: ${variables.size.big};
-  }
-`;
-
-const priority = css`
-  z-index: 2;
-`;
-
-const backGroundColor = css`
+const Priority = css`
   background-color: #fdfcff;
-  margin: -1.8rem;
-  height: 100svh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+  position: relative;
+  z-index: 3;
 `;
-const header = css`
+
+const BackGroundColor = css`
+  background-color: #fdfcff;
+  margin: -2rem -1.8rem;
+`;
+const Header = css`
+  background-color: #fdfcff;
   display: flex;
+  max-width: 50rem;
   width: 100%;
-  min-height: ${variables.headerHeight};
-  justify-content: center;
-  align-items: flex-end;
-  padding: 1.8rem;
+  padding: 2rem 1.8rem;
   box-sizing: border-box;
-  box-shadow: 0 0 1rem rgba(217, 203, 245, 0.37);
+  min-height: ${variables.headerHeight};
+  align-items: flex-end;
+  justify-content: center;
+  box-shadow: 0 0 3rem rgba(217, 203, 245, 0.37);
 
   .prev {
     display: block;
@@ -379,7 +194,7 @@ const header = css`
     background-repeat: no-repeat;
     background-position: center;
     background-size: cotain;
-    width: 2rem;
+    width: 5rem;
     height: 2rem;
   }
 
@@ -398,24 +213,32 @@ const header = css`
   }
 `;
 
-const dateText = css`
-  font-size: ${variables.size.small};
-  color: #868581;
+const ScrollBox = css`
+  max-width: 50rem;
   width: 100%;
-  text-align: center;
-  margin: 2rem auto;
-`;
-
-const messageBox = css`
-  background-color: #fdfcff;
-  width: 100%;
+  height: calc(100svh - ${variables.headerHeight} - 10rem);
+  overflow: hidden auto;
   padding: 1.8rem;
   box-sizing: border-box;
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  & .dateText {
+    font-size: ${variables.size.small};
+    margin-bottom: 2rem;
+    color: #868581;
+    text-align: center;
+  }
+`;
+
+const MessageBox = css`
+  border-radius: 2rem;
   display: flex;
   flex-direction: column;
   gap: 1.4rem;
   align-items: flex-start;
-  border-radius: 2rem;
 
   li > p {
     padding: 1.4rem 2.2rem;
@@ -453,18 +276,18 @@ const messageBox = css`
   }
 `;
 
-const inputBox = css`
+const InputBox = css`
+  background-color: #f8f4ff;
   position: fixed;
   left: 50%;
   bottom: 0;
   transform: translateX(-50%);
-  padding: 1.8rem;
   max-width: 50rem;
   width: 100%;
+  padding: 2rem 1.8rem;
   box-sizing: border-box;
   display: flex;
   gap: 0.8rem;
-  background-color: #f8f4ff;
 
   & input {
     height: 4rem;
